Add files via upload
Here's my p5.js program which creates a square that transforms every frame with a different size and scales with the same color range. It also have a 25% of probability to became a dark version of the program that change the colors with grayscale that has a little bit of color.